Add files via upload
This repository calculates water quality for the Yahara 2070 scenarios. THMB outputs are necessary inputs to the water quality model. The R scripts are run in sequence Step1, Step2, Step3. The .Rdata files are parameter files necessary for the water quality model. The .csv files are outputs from the factorial experiment on the Yahara2070 scenarios. Comments in the R scripts explain the experiment.  The .txt file provides more information.